Move the top-level ProgressIntake() function directly under pkg/graph/resolvers/
